+++ lib/utils/concurrency/taskgroup_wf_test.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"fmt"
 	"math"
-	"math/rand"
 	"strings"
 	"sync"
 	"testing"
@@ -1083,120 +1082,6 @@ func TestChildrenWaitingGameWithTimeoutsButAborting(t *testing.T) {
 			t.FailNow()
 		}
 	}
-}
-
-func TestChildrenWaitingGameWithTimeoutsButAbortingInParallel(t *testing.T) {
-	defer func() { // sometimes this test panics, breaking coverage collection..., so no more panics
-		if r := recover(); r != nil {
-			t.Errorf("Test panicked")
-			t.FailNow()
-		}
-	}()
-
-	failure := false
-	wg := sync.WaitGroup{}
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-
-		overlord, xerr := NewTaskGroup()
-		require.NotNil(t, overlord)
-		require.Nil(t, xerr)
-
-		theID, xerr := overlord.GetID()
-		require.Nil(t, xerr)
-		require.NotEmpty(t, theID)
-
-		fmt.Println("Begin")
-
-		for ind := 0; ind < 100; ind++ {
-			fmt.Println("Iterating...")
-			rint := time.Duration(rand.Intn(20)+30) * 10 * time.Millisecond
-			_, xerr := overlord.Start(func(t Task, parameters TaskParameters) (TaskResult, fail.Error) {
-				delay := parameters.(time.Duration)
-				fmt.Printf("Entering (waiting %v)\n", delay)
-				defer fmt.Println("Exiting")
-
-				dur := delay / 100
-				for i := 0; i < 100; i++ {
-					if t.Aborted() {
-						break
-					}
-					time.Sleep(dur)
-				}
-				return "waiting game", nil
-			}, rint)
-			if xerr != nil {
-				t.Errorf("Unexpected: %s", xerr)
-			}
-		}
-
-		begin := time.Now()
-		go func() {
-			time.Sleep(310 * time.Millisecond)
-			if xerr := overlord.Abort(); xerr != nil {
-				t.Fail()
-			}
-			// did we abort ?
-			aborted := overlord.Aborted()
-			if !aborted {
-				t.Logf("We just aborted without error above..., why Aborted() says it's not ?")
-			}
-		}()
-
-		if _, xerr := overlord.WaitGroup(); xerr != nil {
-			switch xerr.(type) { // nolint
-			case *fail.ErrAborted:
-				// Wanted situation, continue
-			case *fail.ErrorList:
-				el, _ := xerr.(*fail.ErrorList)
-				for _, ae := range el.ToErrorSlice() {
-					if _, ok := ae.(*fail.ErrAborted); !ok {
-						t.Errorf(
-							"everything should be aborts in this test: %v", ae,
-						) // FIXME: CI Failed
-						// This happened on previous CI failure : 3d90f4a9-1e13-446b-881e-002041f02f92: context canceled: context canceled
-						failure = true
-						return
-					}
-				}
-			default:
-				t.Errorf("waitgroup failed with an unexpected error: %v", xerr)
-				failure = true
-				return
-			}
-		} else {
-			t.Errorf("WaitGroup didn't fail and it should")
-			failure = true
-			return
-		}
-
-		end := time.Since(begin)
-
-		fmt.Println("Here we are")
-
-		if end >= (time.Millisecond * 1200) {
-			t.Logf("It should have finished near 1200 ms but it didn't, it was %v !!", end)
-		}
-
-		if end >= (time.Millisecond * 2000) {
-			t.Errorf("It should have finished near 1200 ms but it didn't, it was %v !!", end)
-		}
-	}()
-
-	runOutOfTime := waitTimeout(&wg, 60*time.Second)
-	if runOutOfTime {
-		if failure {
-			t.FailNow()
-		}
-		t.Errorf("Failure: there is a deadlock in TestChildrenWaitingGameWithTimeoutsButAbortingInParallel !")
-		t.FailNow()
-	}
-	if failure {
-		t.FailNow()
-	}
-
-	time.Sleep(3 * time.Second)
 }
 
 func BenchmarkTryWaitGroup(b *testing.B) {
